refactor module with auxiliary functions
