+++ src/utils/common.js
@@ -1,0 +1,4 @@
+function getRandomArrayElement(items) {
+  return items[Math.floor(Math.random() * items.length)];
+}
+export {getRandomArrayElement};
