+++ src/utils/point.js
@@ -3,9 +3,6 @@ import dayjs from 'dayjs';
 const TIME_FORMAT = 'H:m';
 const DATE_FORMAT = 'MMM DD';
 
-function getRandomArrayElement(items) {
-  return items[Math.floor(Math.random() * items.length)];
-}
 
 function humanizePointDueTime(dueDate) {
   return dueDate ? dayjs(dueDate).format(TIME_FORMAT) : '';
@@ -15,5 +12,4 @@ function humanizePointDueDate(dueDate) {
   return dueDate ? dayjs(dueDate).format(DATE_FORMAT) : '';
 }
 
-export {getRandomArrayElement, humanizePointDueDate, humanizePointDueTime};
-
+export {humanizePointDueDate, humanizePointDueTime};
